Refactor `paginate` to handle exception in Bitbucket and Github
`paginate` method is shared between both services and each service
implement `get_next_url_to_paginate` and `get_paginated_results`.
